pyproject.toml
[project]
name = "data-formatter"
version = "0.1.0"
description = "A scalable data formatter for converting between formats and chat stylings using an IR-based approach"
readme = "README.md"
authors = [
    { email = "hjain24in@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
data-formatter = "data_formatter:main"

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"
